Moves all samples into the machine directory
The idea is to use these builder functions in some of the execution
reducers, such as when raising an exception.
