+++ specs/execution/conditional_spec.js
@@ -4,9 +4,9 @@ import {
   sequenceStatement,
   bindingStatement,
   conditionalStatement,
-} from "../samples/statements";
-import { lexicalIdentifier } from "../samples/lexical";
-import { literalNumber, literalRecord } from "../samples/literals";
+} from "../../app/oz/machine/statements";
+import { lexicalIdentifier } from "../../app/oz/machine/lexical";
+import { literalNumber, literalRecord } from "../../app/oz/machine/literals";
 import {
   buildSingleThreadedState,
   buildSemanticStatement,
