+++ specs/execution/exception_catch_spec.js
@@ -3,8 +3,8 @@ import {
   skipStatement,
   exceptionCatchStatement,
   bindingStatement,
-} from "../samples/statements";
-import { lexicalIdentifier } from "../samples/lexical";
+} from "../../app/oz/machine/statements";
+import { lexicalIdentifier } from "../../app/oz/machine/lexical";
 import {
   buildSingleThreadedState,
   buildSemanticStatement,
