+++ specs/execution/exception_context_spec.js
@@ -5,8 +5,8 @@ import {
   exceptionContextStatement,
   exceptionCatchStatement,
   bindingStatement,
-} from "../samples/statements";
-import { lexicalIdentifier } from "../samples/lexical";
+} from "../../app/oz/machine/statements";
+import { lexicalIdentifier } from "../../app/oz/machine/lexical";
 import {
   buildSingleThreadedState,
   buildSemanticStatement,
