+++ specs/free_identifiers/pattern_matching_spec.js
@@ -3,9 +3,9 @@ import { collectFreeIdentifiers } from "../../app/oz/free_identifiers";
 import {
   patternMatchingStatement,
   bindingStatement,
-} from "../samples/statements";
-import { lexicalIdentifier } from "../samples/lexical";
-import { literalRecord } from "../samples/literals";
+} from "../../app/oz/machine/statements";
+import { lexicalIdentifier } from "../../app/oz/machine/lexical";
+import { literalRecord } from "../../app/oz/machine/literals";
 
 describe("Collecting free identifiers in a pattern matching statement", () => {
   beforeEach(() => {
