+++ specs/free_identifiers/skip_spec.js
@@ -1,6 +1,6 @@
 import Immutable from "immutable";
 import { collectFreeIdentifiers } from "../../app/oz/free_identifiers";
-import { skipStatement } from "../samples/statements";
+import { skipStatement } from "../../app/oz/machine/statements";
 
 describe("Collecting free identifiers in a skip statement", () => {
   beforeEach(() => {
