+++ specs/parser/literals/record_spec.js
@@ -1,6 +1,6 @@
 import Immutable from "immutable";
-import { lexicalIdentifier } from "../../samples/lexical";
-import { literalRecord } from "../../samples/literals";
+import { lexicalIdentifier } from "../../../app/oz/machine/lexical";
+import { literalRecord } from "../../../app/oz/machine/literals";
 import { parserFor } from "../../../app/oz/parser";
 import literalGrammar from "../../../app/oz/grammar/literals.ne";
