+++ specs/parser/statements/local_spec.js
@@ -1,6 +1,9 @@
 import Immutable from "immutable";
-import { lexicalIdentifier } from "../../samples/lexical";
-import { skipStatement, localStatement } from "../../samples/statements";
+import { lexicalIdentifier } from "../../../app/oz/machine/lexical";
+import {
+  skipStatement,
+  localStatement,
+} from "../../../app/oz/machine/statements";
 import parse from "../../../app/oz/parser";
 
 describe("Parsing local X in ... end statements", () => {
