+++ specs/parser/statements/procedure_application_spec.js
@@ -1,6 +1,6 @@
 import Immutable from "immutable";
-import { lexicalIdentifier } from "../../samples/lexical";
-import { procedureApplicationStatement } from "../../samples/statements";
+import { lexicalIdentifier } from "../../../app/oz/machine/lexical";
+import { procedureApplicationStatement } from "../../../app/oz/machine/statements";
 import parse from "../../../app/oz/parser";
 
 describe("Parsing {X ...} statements", () => {
