+++ specs/value_creation/record_spec.js
@@ -1,9 +1,9 @@
 import Immutable from "immutable";
 import { createValue } from "../../app/oz/machine/sigma";
 import { buildEnvironment, buildVariable } from "../../app/oz/machine/build";
-import { literalRecord } from "../samples/literals";
-import { valueRecord } from "../samples/values";
-import { lexicalIdentifier } from "../samples/lexical";
+import { literalRecord } from "../../app/oz/machine/literals";
+import { valueRecord } from "../../app/oz/machine/values";
+import { lexicalIdentifier } from "../../app/oz/machine/lexical";
 
 describe("Creating record values in the sigma", () => {
   beforeEach(() => {
